Transferred layer list and legend components
- LayerListComponent
- LayerListCategorizedComponent
- LegendComponent

This causes a circular dependency between layer list components and legend component.

Legend component depends on LayerListService. LayerListComponent depends on LayerListService and LegendService

Need to refactor and possibly rewrite the layer and legend list service implementations to resolve the circular dependency.

This commit includes scratch work using Esri's LayerListViewModel and Accessor property watches.
