+++ apps/aggiemap/src/app/modules/components/sidebar/components/reference/reference.component.html
@@ -1,5 +1,5 @@
 <search [placeholder]="'Find Building or Parking'" [leftActionIcon]="'search'" [rightActionIcon]="'close'" (result)="onSearchResult($event)"></search>
 
-<layer-list></layer-list>
+<tamu-gisc-layer-list></tamu-gisc-layer-list>
 
-<legend></legend>
+<tamu-gisc-legend></tamu-gisc-legend>
