+++ libs/maps/esri/src/lib/services/module-provider.service.ts
@@ -161,6 +161,14 @@ export class EsriModuleProviderService {
     {
       class: 'esri/layers/CSVLayer',
       name: 'CSVLayer'
+    },
+    {
+      class: 'esri/widgets/Legend/LegendViewModel',
+      name: 'LegendViewModel'
+    },
+    {
+      class: 'esri/widgets/LayerList/LayerListViewModel',
+      name: 'LayerListViewModel'
     }
   ];
