+++ libs/maps/feature/layer-list/src/index.ts
@@ -1,3 +1,6 @@
 export * from './lib/maps-feature-layer-list.module';
 
 export * from './lib/services/layer-list.service';
+
+export * from './lib/components/layer-list/layer-list.component';
+export * from './lib/components/layer-list-categorized/layer-list-categorized.component';
