+++ libs/maps/feature/layer-list/src/lib/components/layer-list-categorized/layer-list-categorized.component.html
@@ -1,1 +1,19 @@
-<p>layer-list-categorized works!</p>
+<div class="sidebar-component-name">
+  <span class="material-icons" *ngIf="responsive?.isMobile" (click)="backAction()">arrow_back</span>
+  <p>Layers</p>
+</div>
+
+<div class="sidebar-component-content-container">
+  <div class="category" *ngFor="let category of categorized">
+    <div class="category-title" (click)="category.expanded = !category.expanded">
+      <p>{{ category.title }}</p>
+    </div>
+
+    <!-- TODO: Include later: accordion [accordionExpanded]="category.expanded" -->
+    <div class="category-item">
+      <p *ngFor="let item of category.layers" class="layer-item" [ngClass]="{ loaded: item?.layer?.loaded && item?.layer?.visible }" (click)="toggleLayer(item)">
+        {{ item.title }}
+      </p>
+    </div>
+  </div>
+</div>
