+++ libs/maps/feature/legend/src/index.ts
@@ -1,3 +1,5 @@
 export * from './lib/maps-feature-legend.module';
 
-export * from './lib/services/legend';
+export * from './lib/services/legend.service';
+
+export * from './lib/components/legend/legend.component';
