+++ libs/maps/feature/legend/src/lib/components/legend/legend.component.html
@@ -4,7 +4,7 @@
 </div>
 
 <div class="sidebar-component-content-container">
-  <div class="legend-item" *ngFor="let item of (legend | async)" (click)="analyticsReport(item)">
+  <div class="legend-item" *ngFor="let item of legend | async" (click)="analyticsReport(item)">
     <div class="image-container">
       <img [src]="item.src" [alt]="item.title" />
     </div>
