+++ libs/maps/feature/legend/src/lib/components/legend/legend.component.scss
