+++ libs/maps/feature/legend/src/lib/maps-feature-legend.module.ts
@@ -1,10 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { LegendService } from './services/legend';
+import { LegendService } from './services/legend.service';
+import { LegendComponent } from './components/legend/legend.component';
 
 @NgModule({
   imports: [CommonModule],
-  providers: [LegendService]
+  providers: [LegendService],
+  declarations: [LegendComponent],
+  exports: [LegendComponent]
 })
 export class LegendModule {}
