+++ libs/maps/feature/legend/src/lib/services/legend.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { LegendService } from './legend';
+import { LegendService } from './legend.service';
 
 describe('LegendService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
