+++ libs/maps/feature/legend/src/lib/services/legend.service.ts
